+++ Day1_10082022/WhileLoop/patternNRNC.py
