+++ Day1_10082022/WhileLoop/patternNRNCInt.py
