+++ Day1_10082022/WhileLoop/primeOrNot.py
