+++ Day1_10082022/WhileLoop/sumEvenNoWhile.py
