+++ Day1_10082022/WhileLoop/whileLoopPython.py
