+++ Day32_11082022/FinalSubcs_531/CS531_Inet.txt
@@ -1,0 +1,5 @@
+111.22.3.44 platte
+131.250.95.21 jet
+172.66.7.88 wabash
+111.22.5.66 green
+131.250.47.63 baker
