+++ Day32_11082022/FinalSubcs_531/linkedListQuit.c
@@ -1,0 +1,12 @@
+#include<stdio.h>
+#include<stdlib.h>
+#include<stdbool.h>
+#include<string.h>
+#include "linkedListheader.h"
+/******** Start: Exit File**********/
+void Quit_Function()
+{
+    printf("\nGood bye!");
+    exit(0);
+}
+/******** End: Exit File**********/
